Move torch-based functionality into separate module
