+++ cascade/utils/tests/test_baselines.py
@@ -39,3 +39,4 @@ def test():
         model.predict([0, 0, 0])
         == np.array([[[1, 0], [0, 1]], [[1, 0], [0, 1]], [[1, 0], [0, 1]]])
     )
+
