+++ cascade/utils/torch/__init__.py
@@ -1,0 +1,1 @@
+from .torch_model import TorchModel
